// frontend/src/proto/rider.ts
export interface RideStatus {
    /**
     * @generated from protobuf field: string intent_id = 1
     */
    intentId: string;
    /**
     * @generated from protobuf field: lastmile.rider.RideStatus.Status status = 2
     */
    status: RideStatus_Status;
    /**
     * @generated from protobuf field: string trip_id = 3
     */
    tripId: string;
    /**
     * @generated from protobuf field: string destination_area_id = 4
     */
    destinationAreaId: string;
    /**
     * @generated from protobuf field: string station_area_id = 5
     */
    stationAreaId: string;
}
/**
 * @generated from protobuf enum lastmile.rider.RideStatus.Status
 */
export enum RideStatus_Status {
    /**
     * @generated from protobuf enum value: PENDING = 0;
     */
    PENDING = 0,
    /**
     * @generated from protobuf enum value: SCHEDULED = 1;
     */
    SCHEDULED = 1,
    /**
     * @generated from protobuf enum value: PICKED_UP = 2;
     */
    PICKED_UP = 2,
    /**
     * @generated from protobuf enum value: COMPLETED = 3;
     */
    COMPLETED = 3,
    /**
     * @generated from protobuf enum value: CANCELLED = 4;
     */
    CANCELLED = 4
}
// ...
class RideStatus$Type extends MessageType<RideStatus> {
    constructor() {
        super("lastmile.rider.RideStatus", [
            { no: 1, name: "intent_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "status", kind: "enum", T: () => ["lastmile.rider.RideStatus.Status", RideStatus_Status] },
            { no: 3, name: "trip_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "destination_area_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "station_area_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<RideStatus>): RideStatus {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.intentId = "";
        message.status = 0;
        message.tripId = "";
        message.destinationAreaId = "";
        message.stationAreaId = "";
        if (value !== undefined)
            reflectionMergePartial<RideStatus>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RideStatus): RideStatus {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string intent_id */ 1:
                    message.intentId = reader.string();
                    break;
                case /* lastmile.rider.RideStatus.Status status */ 2:
                    message.status = reader.int32();
                    break;
                case /* string trip_id */ 3:
                    message.tripId = reader.string();
                    break;
                case /* string destination_area_id */ 4:
                    message.destinationAreaId = reader.string();
                    break;
                case /* string station_area_id */ 5:
                    message.stationAreaId = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RideStatus, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string intent_id = 1; */
        if (message.intentId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.intentId);
        /* lastmile.rider.RideStatus.Status status = 2; */
        if (message.status !== 0)
            writer.tag(2, WireType.Varint).int32(message.status);
        /* string trip_id = 3; */
        if (message.tripId !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.tripId);
        /* string destination_area_id = 4; */
        if (message.destinationAreaId !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.destinationAreaId);
        /* string station_area_id = 5; */
        if (message.stationAreaId !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.stationAreaId);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message lastmile.rider.RideStatus
 */
export const RideStatus = new RideStatus$Type();